Token Expiry Issue
In some cases the token check and API call can run over 1 second apart which can cause issues if the token check is at the token expiry time, as the API call fails due to an expired token.

This fixes the issue by renewing the token if there is less than 1 minute of expiry on the token.